zl1: force camera HAL1 for the following apps
Facebook, Facebook Messenger, Instagram, Viber and WhatsApp sometimes
have issues with the camera preview freezing with camera HAL3 on our
devices because the camera blobs were only meant to be used on HAL1,
and on HAL3 sometimes mm-camera-daemon fails to get the preview frames
which causes the preview to freeze.

Forcing these apps to use HAL1 prevents this issue from occurring and
these apps don't require HAL3 for full functionality.

Change-Id: I6926bce825b68f0770c767512372346a479f3ef5